Cross-compilation correctness fixes
In order to build the subpackages that have Python dependencies, they
need to have Python and cross-python as build deps. tensorflow-estimator
further needs to have tensorflow-base as a build dep because its build
process needs to `import tensorflow` in order to generate API bindings,
and in the cross-python setup the build-arch package is required to
provide binary Python modules for import.